Misty-Robot/requirements.txt
anyio==4.2.0
asttokens==2.4.1
cachetools==5.3.2
certifi==2023.11.17
chardet==4.0.0
decorator==5.1.1
exceptiongroup==1.2.0
executing==2.0.1
google-ai-generativelanguage==0.4.0
google-api-core==2.16.2
google-auth==2.27.0
google-cloud-aiplatform==1.42.1
google-cloud-bigquery==3.17.2
google-cloud-core==2.4.1
google-cloud-resource-manager==1.12.2
google-cloud-speech==2.24.0
google-cloud-storage==2.14.0
google-cloud-texttospeech==2.16.3
google-crc32c==1.5.0
google-generativeai==0.3.2
google-resumable-media==2.7.0
googleapis-common-protos==1.62.0
grpc-google-iam-v1==0.13.0
grpcio==1.60.1
grpcio-status==1.60.1
h11==0.14.0
httpcore==1.0.3
httpx==0.25.2
idna==2.10
ipython==8.18.1
jedi==0.19.1
matplotlib-inline==0.1.6
numpy==1.26.4
ollama==0.1.6
packaging==23.2
parso==0.8.3
pexpect==4.9.0
prompt-toolkit==3.0.43
proto-plus==1.23.0
protobuf==4.25.2
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydub==0.25.1
Pygments==2.17.2
python-dateutil==2.8.2
requests==2.25.1
rsa==4.9
shapely==2.0.3
six==1.16.0
sniffio==1.3.0
stack-data==0.6.3
tqdm==4.66.2
traitlets==5.14.1
typing_extensions==4.9.0
urllib3==1.26.18
wcwidth==0.2.13
websocket-client==0.57.0
yapf==0.30.0
